+++ cli-generator/src/main/java/com/dmj/cli/domain/SysRolePermission.java
@@ -1,12 +1,13 @@
 package com.dmj.cli.domain;
 
-import com.baomidou.mybatisplus.annotation.TableName;
 import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableField;
-import java.io.Serializable;
+import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.annotation.TableName;
 import io.swagger.annotations.ApiModel;
 import lombok.Data;
+
+import java.io.Serializable;
 
 /**
  * <p>
@@ -24,12 +25,12 @@ public class SysRolePermission implements Serializable {
     private static final long serialVersionUID=1L;
 
     @TableId(value = "id", type = IdType.ASSIGN_ID)
-    private Integer id;
+    private Long id;
 
     @TableField("role_id")
-    private Integer roleId;
+    private Long roleId;
 
     @TableField("permission_id")
-    private Integer permissionId;
+    private Long permissionId;
 
 }
