+++ cli-generator/src/main/java/com/dmj/cli/domain/UserInfoAccount.java
@@ -1,15 +1,15 @@
 package com.dmj.cli.domain;
 
-import com.baomidou.mybatisplus.annotation.TableName;
 import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableField;
-import java.io.Serializable;
-import java.math.BigDecimal;
-
+import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.annotation.TableName;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
+
+import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * <p>
@@ -27,11 +27,11 @@ public class UserInfoAccount extends BaseEntity implements Serializable {
     private static final long serialVersionUID=1L;
     
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     @ApiModelProperty(value = "用户ID")
     @TableField("user_id")
-    private Integer userId;
+    private Long userId;
 
     @ApiModelProperty(value = "账户类型(VIDEO:观看账户 DOWNLOAD:下载账户)")
     @TableField("account_type")
