+++ cli-generator/src/main/java/com/dmj/cli/mapper/sys/SysUserMapper.java
@@ -21,7 +21,7 @@ public interface SysUserMapper extends BaseMapper<SysUser> {
 
     SysUserDTO getUserByName(@Param("userName") String userName);
 
-    List<Long> listMenuIdsByUserId(@Param("userId") Integer userId);
+    List<Long> listMenuIdsByUserId(@Param("userId") Long userId);
 
     List<SysUserVO> listSysUser(UserQuery userQuery);
 }
